+++ interface/MultiDimFit.h
@@ -12,6 +12,7 @@
 #include <RooRealVar.h>
 #include "TFile.h"
 #include <vector>
+#include <map>
 #include <TFile.h>
 
 class MultiDimFit : public FitterAlgoBase {
@@ -35,7 +36,7 @@ protected:
 
   static std::vector<std::string>  poi_;
   static std::vector<RooRealVar*>  poiVars_;
-  static std::vector<double>       poiVals_;
+  static std::vector<float>       poiVals_;
   static RooArgList                poiList_;
   static unsigned int              nOtherFloatingPoi_; // keep a count of other POIs that we're ignoring, for proper chisquare normalization
   static double                    deltaNLL_;
@@ -77,7 +78,7 @@ protected:
   static std::string saveSpecifiedIndex_;
   static std::vector<std::string>  specifiedFuncNames_;
   static std::vector<RooAbsReal*> specifiedFunc_;
-  static std::vector<double>       specifiedFuncVals_;
+  static std::vector<float>       specifiedFuncVals_;
   static RooArgList                specifiedFuncList_;
   static std::vector<std::string>  specifiedCatNames_;
   static std::vector<RooCategory*> specifiedCat_;
@@ -85,7 +86,7 @@ protected:
   static RooArgList                specifiedCatList_;
   static std::vector<std::string>  specifiedNuis_;
   static std::vector<RooRealVar *> specifiedVars_;
-  static std::vector<double>       specifiedVals_;
+  static std::vector<float>       specifiedVals_;
   static RooArgList                specifiedList_;
   static bool saveInactivePOI_;
   static bool skipDefaultStart_;
@@ -101,7 +102,7 @@ protected:
   void doStitch2D(RooWorkspace *w, RooAbsReal &nll) ;
   void doImpact(RooFitResult &res, RooAbsReal &nll) ;
 
-  std::map<std::string, std::vector<double>> getRangesDictFromInString(std::string) ;
+  std::map<std::string, std::vector<float>> getRangesDictFromInString(std::string) ;
 
 
   // utilities
